Refactor pagination to be independent react component
